+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0b47 (Unreleased)
 
 ### Features Added
-- Rename metric names for customer sdk stats and set it on by default
+- Add custom metric mapping for customer sdkstats metric names to preserve casing
   ([#44849](https://github.com/Azure/azure-sdk-for-python/pull/44849))
 - Add auto detection for application ID from connection string if not set 
   ([#44644](https://github.com/Azure/azure-sdk-for-python/pull/44644))